requirements.txt
appdirs==1.4.4
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.9.3
boto3==1.18.22
botocore==1.21.22
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.8.0
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
fake-useragent==0.1.11
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.6.4
jmespath==0.10.0
lxml==4.6.3
Markdown==3.3.4
parse==1.19.0
Pillow==8.3.1
playwright==1.14.0
psycopg2==2.9.1
pycodestyle==2.7.0
pyee==8.2.2
PyJWT==2.1.0
pyppeteer==0.2.6
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt6==6.1.1
PyQt6-Qt6==6.1.2
PyQt6-sip==13.1.0
pyquery==1.4.3
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.1
requests==2.26.0
requests-html==0.10.0
s3transfer==0.5.0
selenium==3.141.0
selenium-stealth==1.0.6
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
TikTokApi==4.0.1
toml==0.10.2
tqdm==4.62.1
typing-extensions==3.10.0.0
urllib3==1.26.6
w3lib==1.22.0
websockets==9.1
whitenoise==5.3.0
zipp==3.5.0
